refactor(tests): rename test files
