+++ test/logger/set-logger.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { setLogger } from "../../src/logger/index.js";
+import { setLogger } from "../../src/logger/set-logger.js";
 
 import { PACKAGE_NAME } from "../../src/shared/constants.js";
